fix(ci): repair failing checks after pi-ai 0.84.3 bump
Two CI failures on main after the pi-* 0.84.3 bump (#852/#853):

1. Mobile — Android / bundle budget: main entry hit 505.10 KB gzip vs the
   505 KB budget. Root cause: TriggersPanel.tsx (2700 lines, React.lazy in
   App.tsx) was statically imported for getIncompleteTriggers by
   slash-commands.ts and useTriggerCount.ts, baking the whole panel into
   the main chunk and defeating the lazy split. Extracted the pure
   grouping helpers into attention/trigger-groups.ts; TriggersPanel now
   code-splits properly. Main entry drops 503.5 → 491.27 KB gzip locally.
   Logic tests moved to attention/trigger-groups.test.ts — no more global
   UI mocks (the old mock.module scaffolding leaked into other test files).

2. Unit Tests / nvidia-models.test.ts: pi-ai 0.84.3 curates
   moonshotai/kimi-k3 in the NVIDIA catalog, so the 'never seen' test id
   now resolves to curated metadata, and the new per-model
   thinkingLevelMaps (deepseek entries null out low/medium) degraded the
   thinking control to [off, high]. The overlay now forces NVIDIA's
   accepted values (low|medium|high) on top of any upstream map, and the
   tests use a synthetic unknown id.
