+++ packages/ui/src/attention/trigger-groups.test.ts
@@ -1,46 +1,15 @@
 /**
- * Pure-logic tests for TriggersPanel exported helpers.
+ * Pure-logic tests for trigger grouping helpers.
  *
- * These test groupByLinkedSession, isPendingTrigger, and getIncompleteTriggers
- * without any DOM, React, or module-alias dependencies.
+ * Tests groupByLinkedSession, isPendingTrigger, and getIncompleteTriggers
+ * without any DOM, React, or module-alias dependencies. These helpers used
+ * to live inside TriggersPanel.tsx (requiring global UI mocks — see git
+ * history); extracting them made the mocks unnecessary.
  */
-import { describe, test, expect, mock } from "bun:test";
-
-// ── Minimal mocks for TriggersPanel's UI imports ─────────────────────────────
-// IMPORTANT: bun's mock.module is process-global and is NOT undone by
-// mock.restore(), so these leak into every test file that runs after this one.
-// The stubs must therefore render their children (a `() => null` Button here
-// blanked out buttons in unrelated component tests, e.g. DeviceSetupScanner's
-// "Allow Camera & Scan" — which tests failed depended on file order).
-/* eslint-disable @typescript-eslint/no-explicit-any */
-const R = await import("react");
-const childStub = (tag: string) => {
-  const Stub = R.forwardRef(({ children, ...props }: any, ref: any) =>
-    R.createElement(tag, { ...props, ref }, children),
-  );
-  Stub.displayName = `Stub(${tag})`;
-  return Stub;
-};
-mock.module("@/components/ui/button", () => ({ Button: childStub("button") }));
-mock.module("@/components/ui/badge", () => ({ Badge: childStub("span") }));
-// Dialog stubs stay null: real dialogs hide content until opened, so a
-// children-rendering stub would leak inline dialog text into later files
-// and break their single-match queries the same way.
-mock.module("@/components/ui/dialog", () => ({
-  Dialog: () => null, DialogContent: () => null, DialogHeader: () => null,
-  DialogTitle: () => null, DialogFooter: () => null, DialogDescription: () => null,
-}));
-mock.module("@/lib/utils", () => ({ cn: (...args: any[]) => args.filter(Boolean).join(" ") }));
-/* eslint-enable @typescript-eslint/no-explicit-any */
-
-// Import AFTER mocks are registered
-const {
-  groupByLinkedSession,
-  isPendingTrigger,
-  getIncompleteTriggers,
-  RESPONSE_TRIGGER_TYPES,
-} = await import("./TriggersPanel");
-import type { TriggerHistoryEntry } from "./TriggersPanel";
+import { describe, test, expect } from "bun:test";
+import type { TriggerHistoryEntry } from "./trigger-utils";
+import { isPendingTrigger, RESPONSE_TRIGGER_TYPES } from "./trigger-utils";
+import { groupByLinkedSession, getIncompleteTriggers } from "./trigger-groups";
 
 function makeTrigger(overrides: Partial<TriggerHistoryEntry> = {}): TriggerHistoryEntry {
   return {
